restore test_prophesee and fix lint
